Switch to JDK 9
Both @PostConstruct and @PreDestroy annotations (here being used in net.will.javatest.javax.annotation.Irrelevant) are located in java.xml.ws.annotation package which is a part of  Java EE module. J2EE has been deprecated in Java 9 and it’s scheduled to be removed in Java 11.

The solution to me for now is adding the necessary dependency to build.gradle. Another possible solution in the future could be just removing the tests using these 2 annotations, as they are not really relevant.